shot_chart module and tests
